requirements.txt
agno
#openai
python-dotenv
duckduckgo-search
pypdf
groq 
yfinance 